change double by unsigned long in config files

+++ src/Q4S-client-server/Q4SAgent/Q4SAgentConfigFile.cpp
@@ -9,6 +9,6 @@ Q4SAgentConfigFile::Q4SAgentConfigFile(const std::string &fName)
     ganyConnectorIp = configFile.getValueOfKey<std::string>("GANY_CONNECTOR_IP", "127.0.0.1");
     ganyConnectorPort = configFile.getValueOfKey<std::string>("GANY_CONNECTOR_PORT", "27015");
     listenUDPPort = configFile.getValueOfKey<std::string>("LISTEN_UDP_PORT", "27017");
-    timeEndApp = configFile.getValueOfKey<double>("TIME_END_APP", 10000);
+    timeEndApp = configFile.getValueOfKey<unsigned long>("TIME_END_APP", 10000);
     demoMode= configFile.keyExists("DEMO_MODE");
 }
